requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
corsheaders>=4.0

reportlab>=4.0
matplotlib>=3.8

Pillow>=10.0
python-dateutil>=2.8

# PostgreSQL adapter
psycopg2-binary>=2.9

# Environment variable support
python-dotenv>=1.0